backend/requirements.txt
asgiref==3.3.4
django==3.2
djangorestframework>=3.12.2,<3.13.0


